[Bronze V] Title: 이상한 기호, Time: 124 ms, Memory: 14224 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -9,8 +9,8 @@ public class Main {
 		
 		StringTokenizer st = new StringTokenizer(in.readLine());
 		
-		int A = Integer.parseInt(st.nextToken());
-		int B = Integer.parseInt(st.nextToken());
+		long A = Long.parseLong(st.nextToken());
+		long B = Long.parseLong(st.nextToken());
 		
 		System.out.println((A + B) * (A - B));
 	}
